requirements.txt
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Deprecated==1.2.14
Flask==3.0.3
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.1.3
pandas==2.2.3
pycparser==2.22
PyGithub==2.5.0
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rich==13.9.4
six==1.16.0
smmap==5.0.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.2
wrapt==1.16.0
